pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "persisto"
version = "0.1.0"
description = "Python SDK for the Persisto semantic memory API"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Proprietary" }
authors = [{ name = "Antler Labs", email = "team@antlerlabs.ai" }]
dependencies = [
  "httpx>=0.26",
  "pydantic>=2.5",
  "tenacity>=8.2",
  "python-dotenv>=1.0",
]

[tool.setuptools]
packages = ["persisto"]
